fix(sm12x): int64 block offsets in the non-rowwise paged-MQA-logits kernel too
197d21ee79 cast block_idx to int64 in _fp8_paged_mqa_logits_rowwise_kernel but
left _fp8_paged_mqa_logits_kernel (the non-rowwise variant) unfixed: its
`block_idx * stride_sb` (scale) and `block_idx[:, :, None] * stride_kvb` (KV)
still overflow int32 for the post-rebase packed-KV block stride (~1039680) at
higher block ids. conc=1 exercises the rowwise kernel (already clean) but conc=8
hits this one and hard-crashes (Xid 31 MMU fault, SM121 / IMA on SM120). Cast
both block-offset multiplies to int64 to match the rowwise kernel.

Validated on GB10 2-node (SM121, MTP2): conc=8 long-context coherence 68/68
(0 gibberish, 0 recall-miss; was a hard CUDA crash), GSM8K-200 0.965.

+++ vllm/models/deepseek_v4/nvidia/ops/sm12x_mqa.py
@@ -249,7 +249,7 @@ def _fp8_paged_mqa_logits_kernel(
 
     logits = tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32)
     scale = tl.load(
-        scale_ptr + block_idx * stride_sb + block_offset[None, :] * stride_ss,
+        scale_ptr + block_idx.to(tl.int64) * stride_sb + block_offset[None, :] * stride_ss,
         mask=context_mask,
         other=0.0,
     )
@@ -268,7 +268,7 @@ def _fp8_paged_mqa_logits_kernel(
             ).to(tl.float32)
             k = tl.load(
                 kv_ptr
-                + block_idx[:, :, None] * stride_kvb
+                + block_idx[:, :, None].to(tl.int64) * stride_kvb
                 + block_offset[None, :, None] * stride_kvs
                 + d[None, None, :] * stride_kvd,
                 mask=context_mask[:, :, None] & (d[None, None, :] < head_dim),
